We use your RXTX library (javax.comm) and everything worked well untill we moved from RedHat 6.2 to SuSE 7.2. RXTX started to complain about unexpected lock files.
After investigating the problem we saw that the code in fhs_lock that searches
for these
files, does not take symlinks into account. However, on a SuSE 7.x system the
'/usr/spool/locks'
directory is linked to the '/var/lock' directory, and '/var/lock' is the
'normal' place for lock files.

If a process creates a lock file, and dies, this file is found in two places,
and reported as such.
The fhs_lock tries to skip '/var/lock', but that does not work because it does
only strncmp to check.

The patch included modifies this code. It now obtains the 'stat' of the other
directory, and compares
the st_dev and st_ino fields. As far as I know, these fields together should be
unique for any directory
or file in the system.

regards,

Peter Beelen
Philips TASS