Subject: [Rxtx] patch - omitted unlock on open failures for Mac OS X port
Hello-

I was using the rxtx package with Mac OS X with bluetooth mounted to
the serial port.  The problem I had is that whenever the bluetooth
device was unavailable, the rxtx package would correctly fail to
connect, but incorrectly held on to the lock for the serial port.  This
prevented connection later when the bluetooth device became available.
  I fixed this by releasing the lock on the connection error.  Note: I
have not committed this change to CVS.

-Tico